leaf: Dynamic bootanimation colors
To archive this you have to give your bootanimation solid colors.
In this bootanimation I went for #FF0000 for red
and #00FF00 for green.
In the desc.txt you can see a new line.
The first hex color is for changing #FF0000 values, the following
is for changing #00FF00 values and the third one for #0000FF values.
The fourth value is for alpha channel, which we need to set to BLACK
here so the black pixels dont get another color.
Now if we boot, we gonna see the normal leaf bootanimation, since I
replaced the colors with the same ones.
But if we now add the dynamic color feature to android, the
bootanimation should shance accordingly to your material theme!

Change-Id: Ib4eb63cf9a1b982a3c0c46990eedc44df27b3ef2